requirements.txt
numpy
scipy
pyside6
polars
superqt
watchdog
pyarrow
natsort
pyopengl
